requirements.txt
Django==4.2.10
python-dotenv==1.0.0
django-allauth==0.57.0
Pillow==9.5.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
stripe==7.8.1
django-environ==0.10.0
whitenoise==6.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.3 